api/requirements.txt
alembic==1.11.1
fastapi==0.97.0
fastapi-camelcase==1.0.5
SQLAlchemy==2.0.16
sqlalchemy_utils==0.41.1
psycopg2==2.9.6
uvicorn==0.22.0
PyJWT==2.7.0
passlib==1.7.4
python-multipart==0.0.6
pydantic[email]
httpx==0.24.1